rename quay-org to quay-membership
