+++ reconcile/cli.py
@@ -7,14 +7,14 @@ import reconcile.config as config
 import reconcile.gql as gql
 import reconcile.github_org
 import reconcile.openshift_rolebinding
-import reconcile.quay_org
+import reconcile.quay_membership
 
 from reconcile.aggregated_list import RunnerException
 
 services = {
     'github': reconcile.github_org,
     'openshift-rolebinding': reconcile.openshift_rolebinding,
-    'quay-org': reconcile.quay_org,
+    'quay-membership': reconcile.quay_membership,
 }
